disptacher/requirements.txt
amqp==5.2.0
anyjson==0.3.3
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==5.0
django-bulk-update==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
dnspython==2.4.2
drf-yasg==1.21.7
eventlet==0.33.3
flower==2.0.1
greenlet==3.0.1
gunicorn==21.2.0
humanize==4.9.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jose==1.0.0
kombu==5.3.4
MarkupSafe==2.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.2
prometheus-client==0.19.0
prompt-toolkit==3.0.41
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
simplejson==3.19.2
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
tornado==6.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
